Reduce crit scanner maintenance debt
Split catalog-backed weapon mount counting and derived crit-system detection out of the BV crit scanner so the scanner no longer trips the high file-bloat gate while keeping the remaining advisory split tracked in the maintenance ledger.

Constraint: scripts/** is ignored, so the new intentional validator helpers are force-added as tracked script modules.

Rejected: Ledger-only update | would hide the high finding without reducing scanner size or preserving a reviewable helper boundary.

Confidence: high

Scope-risk: narrow

Directive: Keep future validate-bv scanner splits guarded by validate:bv parity and maintenance-ledger validation.

Tested: node scripts/qc/validate-maintenance-warning-ledger.mjs; npm run validate:bv; npm run typecheck -- --pretty false; npm run qc:validate; npm run qc:app-completion -- --json; npm run maintain:scan:gate; git diff --check; git diff --cached --check

Not-tested: Full repo-wide legacy maintenance debt burn-down beyond this tracked scanner slice.
